added cond if prev value of key is the same + some fixes
